requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
requests
python-dotenv
slack-sdk==3.33.4
pydantic
pypdf
git+https://github.com/rahul-08-11/pyzohobooks.git@main#egg=pyzoho-books
git+https://github.com/rahul-08-11/pyzohocrm.git@main#egg=pyzohocrm